+++ app/src/main/java/com/challengeonair/challengeonairandroid/view/home/ChildAdapter.kt
@@ -6,8 +6,8 @@ import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.challengeonair.challengeonairandroid.R
-import com.challengeonair.challengeonairandroid.model.data.Challenge
-import com.challengeonair.challengeonairandroid.model.data.Category
+import com.challengeonair.challengeonairandroid.model.data.entity.Challenge
+import com.challengeonair.challengeonairandroid.model.data.entity.Category
 import com.challengeonair.challengeonairandroid.databinding.ItemCategoryChallengeBinding
 
 class ChildAdapter(
