+++ app/src/main/java/com/challengeonair/challengeonairandroid/view/home/HomeActivity.kt
@@ -7,8 +7,8 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.challengeonair.challengeonairandroid.databinding.ActivityHomeBinding
-import com.challengeonair.challengeonairandroid.model.data.Category
-import com.challengeonair.challengeonairandroid.model.data.Challenge
+import com.challengeonair.challengeonairandroid.model.data.entity.Category
+import com.challengeonair.challengeonairandroid.model.data.entity.Challenge
 import com.challengeonair.challengeonairandroid.view.alarm.AlarmActivity
 import com.challengeonair.challengeonairandroid.view.challenge.CreateChallengeActivity
 import com.challengeonair.challengeonairandroid.view.mypage.MyPageActivity
