+++ app/src/main/java/com/challengeonair/challengeonairandroid/view/mypage/WaitingChallengeAdapter.kt
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.challengeonair.challengeonairandroid.R
 import com.challengeonair.challengeonairandroid.databinding.ItemWaitingChallengeBinding
-import com.challengeonair.challengeonairandroid.model.data.Challenge
+import com.challengeonair.challengeonairandroid.model.data.entity.Challenge
 
 class WaitingChallengeAdapter(
     private val waitingChallengeList: List<Challenge>,
