+++ app/src/main/res/layout/item_category_challenge.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="challenge"
-            type="com.challengeonair.challengeonairandroid.model.data.Challenge" />
+            type="com.challengeonair.challengeonairandroid.model.data.entity.Challenge" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
