+++ app/src/main/res/layout/item_waiting_challenge.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="challenge"
-            type="com.challengeonair.challengeonairandroid.model.data.Challenge" />
+            type="com.challengeonair.challengeonairandroid.model.data.entity.Challenge" />
 
     </data>
